# Conflicts:
#	jdevelops-data-es-demo/src/main/java/com/example/jdevelopsdataesdemo/entity/TestResArticleES.java
#	jdevelops-data-es-demo/src/main/java/com/example/jdevelopsdataesdemo/entity/TestResArticleESNestedChild2.java

+++ jdevelops-data-es-demo/src/main/java/com/example/jdevelopsdataesdemo/entity/IgnoreResArticleES.java
@@ -29,7 +29,7 @@ import static cn.jdevelops.annotation.es.constant.EsTypeDataFormat.*;
 public class IgnoreResArticleES {
 
 
-    @EsField(basic = @EsFieldBasic(type = EsType.integer, index = false))
+    @EsField(basic = @EsFieldBasic(type = EsType.LONG, index = false))
     private Long id;
 
     /**
